requirements.txt
google-cloud-aiplatform
transformers
torch
